Move object.Object into the code package.
